### 2021-12-21 #4020 @whedon
[PRE REVIEW]: NeuralHydrology --- A Python library for Deep Learning research in hydrology
**Submitting author:** @kratzert (<a href="http://orcid.org/0000-0002-8897-7689">Frederik Kratzert</a>)
**Repository:** <a href="https://github.com/neuralhydrology/neuralhydrology" target ="_blank">https://github.com/neuralhydrology/neuralhydrology</a>
**Version:** v.1.1.0
