### 2021-12-21 #4020 @kratzert
Thanks @arfon was just about to do the same. I haven't seen an option to add an alternative branch while submitting the paper.
