### 2021-12-22 #4020 @elbeejay
@kratzert I am going to poke around in the repository and the code over the next few days before we sent this out for review, but in the meantime, do you mind suggesting some potential reviewers?

Per the post up top, JOSS does have a [list](https://bit.ly/joss-reviewers) of folks who have indicated they are willing to review submissions, so that may be a good starting point (although that is certainly not a requirement for potential reviewers). When you mention potential reviewers please do so **without** tagging them with an @ so that we do not unnecessarily ping tons of people. Thanks!

Feel free to reach out if you have any questions,
Jay
