### 2021-12-22 #4020 @kratzert
Hi @elbeejay

thank you for serving as editor.

Here are some reviewer suggestions from the JOSS reviewer list:
- Andrew Bennett (arbennett): Actually, I think he is/was even working with our library, so he might have some comments/insights from the user perspective already.
- Daniel Buscombe (dbuscombe-usgs): from his fields of expertise, he should be a good match, as he lists hydrology and deep learning
-  Raoul Collenteur (raoulcollenteur): hydrologist with good Python skills.

Additional suggestions:
- Joe Hamman (jhamman): One of the xarray dev's and Pangeo contributors and I know he has been using our library as well, so again, maybe a good candidate for feedback from the user/dev perspective.
- Stephan Hoyer (shoyer): Contributer/Dev to various open source projects, such as Xarray, JAX, TensorFlow etc. I have a lot of respect for his software development skills and wanted to suggest him. But since I joined Google recently, we now share the same affiliation which probably eliminates him as a reviewer.

Let me know if I should think of more reviewer candidates and again, thanks for serving as editor.

Enjoy the holidays.
