### 2021-12-26 #4020 @elbeejay
@kratzert,

Thank you for the suggestions. I've had a chance to take a look at the package and have a few questions and several suggestions (mostly documentation related) for you before we send this out for review.

- What are the system requirements for someone to train, test, and apply models using this package? I'm guessing a nvidia GPU is required although the documentation does not make this clear. I'd like to clear this question up before reaching out to reviewers as I don't want to get stuck in a situation where we have reviewers who are unable to easily apply and test the package.
- In a similar vein, I think it would be helpful to specify the Python version(s) that are compatible with the `neuralhydrology` package, via text in the documentation or a `requirements.txt` file could be good for this
- This should come up in the review process, but I'll note it now, I didn't see any documentation or guidance on how someone would contribute to the package (this is a [JOSS review item](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines))
- I think the documentation would be helped by connecting the hosted documentation to the `examples/` subdirectory more directly. Right now the existence of those configuration and notebook files is not mentioned in the documentation (as far as I could see), and they could help someone who is trying to work through the tutorials
- As provided, I think those example notebooks themselves could be cleaned up a little to remove "hard paths" to files on developer's local machines in favor of relative paths. Similarly some of the instructions for how to get the data and where to save it such that it meshes with the instructions in the configuration file are not provided in the notebooks themselves.

The key question is really the first one, as that will help us find reviewers who are able to test and review this submission. Subsequent points are not critical at this time, however I will note that clear documentation and easy to run example cases and tutorials often makes the review process go much more smoothly.

Thanks,
Jay
