### 2021-12-29 #4020 @kratzert
@elbeejay thanks for your quick comments and sorry for my delayed reply:

1) No special system requirements for using our package. Training one of the models on any more or less modern pc/laptop should work and complete within (less than a) minute(s). Sure, training larger models on larger datasets (which we do mostly) benefits heavily from an CUDA-capable GPU but I think that is no requirement for reviewing the package.

2) Good point. We do provide environment files for conda but somehow missed to make this clearer in the documentation. We already have an update in our private development repository that will be pushed to public shortly. For the meantime you can find the environment files [here](https://github.com/neuralhydrology/neuralhydrology/tree/master/environments), one for a CPU-only environment and two for different CUDA environments. Minimum Python version should be 3.6, I think. We state 3.7 in our environment files but I think this is just an artifact from when we created those files. I will definitely check what are the minimum requirements.

3) Another good point. We will add a section to the documentation on how to contribute.

4) Again good point, thanks. We will edit the examples to include this information at the very top so that it is clearer from the compiled documentation.

5) Thanks, I will work on that.

I am currently in baby bonding until the second week of January. I won't have much time to work on this until then, since we also have our older son at home all day, because his kindergarten is closed. I hope that is okay.

All the best for the last days of the year.
