### 2021-12-30 #4020 @elbeejay
:wave: @dbuscombe-usgs, @jhamman, and @chuckaustin would you be interested in reviewing this submission, "NeuralHydrology --- A Python library for Deep Learning research in hydrology," for JOSS? My understanding is that you all have experience working with either Python or deep learning (or both). At JOSS we have an open checklist-driven review process that takes place through issues on GitHub; review criteria can be found here: https://joss.readthedocs.io/en/latest/review_criteria.html

This is a pre-review issue which we use to find reviewers for a given submission. Once 3 reviewers are found we will formally start the review in a dedicated GitHub issue. At present we are asking reviewers to complete their review of a submission within 6 weeks of the formal review issue opening, although this timeline can be extended as needed.

I'll note that for this submission in particular, I've heard enough about folks in the community using this package that I do not think we will have trouble finding other reviewers if you are unable or uncomfortable with serving as a reviewer. So that being said, please do not feel any pressure to accept this review request; if you are unable to review this and have someone else who you think would be a great reviewer for this package, please mention their GitHub handle here (without the "@" symbol).

Otherwise, if you are interested in reviewing, please just take a look at the JOSS [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html?#joss-conflict-of-interest-policy) before accepting this invitation to ensure we don't run into issues there.

Thank you, feel free to reach out if you have any questions about the JOSS review process. Please do not feel any pressure to accept this review request if you do not have the time or do not feel comfortable reviewing this software package, we appreciate and respect our peer-reviewers' time.

-Jay
