### 2021-12-31 #4020 @dbuscombe-usgs
Thanks this looks interesting, but I'm on vacation now and a little slammed when I'm back so I'll skip, but I should be able to help next time
