### 2022-01-03 #4020 @elbeejay
@ammilten please confirm that you are able and willing to review this submission for JOSS. Thank you!
