+++ CONTRIBUTING.rst
@@ -1,0 +1,103 @@
+Contributing to NeuralHydrology
+===============================
+
+Welcome to the NeuralHydrology contribution guide!
+--------------------------------------------------
+
+First off, thank you for considering contributing to NeuralHydrology!
+
+The following is a set of guidelines for contributing to NeuralHydrology. They are intended to make contributions as easy as possible for both you as a contributor and us as maintainers of the project.
+
+
+Kinds of contribution
+---------------------
+
+There are many ways to contribute to this project. To name a few:
+
+- File bug reports or feature requests as GitHub issues
+- Contribute a bug fix as a pull request
+- Contribute a new feature (new model, new dataset, etc.) as a pull request
+- Improve the `documentation <https://neuralhydrology.readthedocs.io/>`__ with new tutorials, better descriptions, or even just by fixing typos
+
+The following sections will give some more guidance for each of these options.
+
+Reporting Bugs and feature requests
+-----------------------------------
+If you've come across some behavior that you think is a bug or a missing feature, the first step is to check if it's already known.
+For this, take a look at the `GitHub issues page <https://github.com/neuralhydrology/neuralhydrology/issues>`__.
+
+If you can't find anything related there, you're welcome to create your own issue.
+
+**NOTE**: Please use the `issues page <https://github.com/neuralhydrology/neuralhydrology/issues>`__ for bug reports and feature requests, and leave the `discussions <https://github.com/neuralhydrology/neuralhydrology/discussions>`__ for more open discussions, e.g., project ideas or engaging with other users.
+
+
+Creating an issue for a bug report
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+If you couldn't find an open issue addressing the problem, open a new one. Please make sure to use a meaningful title and a clear description that includes as much relevant details as possible. Ideally, a bug report should contain:
+
+- A description of how to reproduce the bug (which commands did you execute, did you do any modifications to the code, etc.)
+- The full stack trace if any exceptions occurred
+- The configuration yaml file you used in your experiments.
+- The NeuralHydrology version you used in your experiments (the git commit if you cloned the repo, or the version number in ``neuralhydrology/__about__.py``)
+- A description of the dataset you are using in your experiments: Are you using an out-of-the-box CAMELS dataset, or did you create your own csv/netCDF files?
+- A list of Python packages you have installed in your environment. If you're using conda, you can create this list via ``conda env export``.
+
+For console output and config files (stack traces, environment files, ...), please try to format them as code directly in the issue (using GitHub's `code block <https://docs.github.com/en/github/writing-on-github/working-with-advanced-formatting/creating-and-highlighting-code-blocks>`__ syntax) or as a link to a `gist <https://gist.github.com/discover>`__.
+
+
+Contributing code with a pull request
+-------------------------------------
+If you wrote some code to fix a bug or provide some new functionality, we use use pull requests (PRs) to merge these changes into the project.
+
+In a PR, please provide a small description of the changes you made, including a reference to the issue that your PR solves (if one exists).
+We'll take a look at your PR and do a small code review, where we might outline changes that are necessary before we can merge your PR into the project.
+Please don't feel overwhelmed by our change requests---it's entirely normal to have a code review go back and forth a few times before the code is ready to be merged.
+
+Once everyone is happy with the changes in the PR, we'll approve and merge it into the master branch.
+
+
+Environment setup
+~~~~~~~~~~~~~~~~~
+
+The `quick start section <https://neuralhydrology.readthedocs.io/en/latest/usage/quickstart.html>`__ contains information on how to set up the environment (installing the required dependencies and the package itself, as well as downloading data).
+
[…]
