### 2022-01-05 1505c83
Add option to save git diff in run_dir (#114)
Add `save_git_diff` config option to save code changes in a diff file
inside the run dir.
