### 2022-01-09 #4020 @elbeejay
:wave: @jhamman and @chuckaustin just trying to follow-up here. Will be pinging you both via email as well.

@raoulcollenteur would you be at all interested in reviewing this submission? No pressure as I see you've relatively recently reviewed another paper for JOSS.
