### 2022-01-11 #4050 @whedon
[REVIEW]: NeuralHydrology --- A Python library for Deep Learning research in hydrology
**Submitting author:** @kratzert (<a href="http://orcid.org/0000-0002-8897-7689">Frederik Kratzert</a>)
**Repository:** https://github.com/neuralhydrology/neuralhydrology
**Branch with paper.md** (empty if default branch): paper
**Version:** v.1.2.3
