### 2022-01-11 #4050 @elbeejay
@kratzert - if looks like there's an issue with the bibliography file. I suspect it has to do with an "@" symbol in the citation:

> @article{kratzert@regional,
> author = {Kratzert, F. and Klotz, D. and Shalev, G. and Klambauer, G. and Hochreiter, S. and Nearing, G.},
> title = {Towards learning universal, regional, and local hydrological behaviors via machine learning applied to large-sample datasets},
> journal = {Hydrology and Earth System Sciences},
> volume = {23},
> year = {2019},
> number = {12},
> pages = {5089--5110},
> doi = {10.5194/hess-23-5089-2019}
> }

Can you remove the second "@" and update the bib file? Once you've done that please use the whedon commands like above to regenerate the paper and check the reference file.
