+++ paper.md
@@ -32,7 +32,7 @@ bibliography: bibliography.bib
 # Summary and statement of need
 
 Since ancient times humans have strived to describe environmental processes related to water [@biswas1970history; @angelakis2012evolution].
-Throughout this history, hydrologists built various process-based prediction models that simulate processes from soil moisture to streamflow generation (a collection of historical references can be found in @loague2010rainfall). More recently, Deep Learning models have emerged as extremely powerful and more accurate alternatives to these traditional modeling approaches [@kratzert@regional; @kratzert2019pub; @gauch2021mts; @klotz2021uncertainty; @lees2021benchmarking]. For hydrologists, embracing this new data-driven paradigm is challenging [@nearing2021role; @keith2020deep]: Not only are the models conceptually different, but they are also built, optimized, and evaluated with different strategies and toolsets.
+Throughout this history, hydrologists built various process-based prediction models that simulate processes from soil moisture to streamflow generation (a collection of historical references can be found in @loague2010rainfall). More recently, Deep Learning models have emerged as extremely powerful and more accurate alternatives to these traditional modeling approaches [@kratzert2019regional; @kratzert2019pub; @gauch2021mts; @klotz2021uncertainty; @lees2021benchmarking]. For hydrologists, embracing this new data-driven paradigm is challenging [@nearing2021role; @keith2020deep]: Not only are the models conceptually different, but they are also built, optimized, and evaluated with different strategies and toolsets.
 
 NeuralHydrology is a Python library based on PyTorch [@paszke2019pytorch] that is designed to build, apply, and experiment with Deep Learning models with a strong focus on hydrological applications. Originally designed for our internal research needs, the library was generalized and open-sourced to allow anyone to experiment with Deep Learning models as easily as possible: Pre-built models and data loaders allow for quick experiments, yet the framework is also easily extensible to new models, data sets, loss functions, or metrics to suit more advanced use-cases.
 
