+++ bibliography.bib
@@ -23,7 +23,7 @@
   address={Wallingford, UK}
 }
 
-@article{kratzert@regional,
+@article{kratzert2019regional,
 author = {Kratzert, F. and Klotz, D. and Shalev, G. and Klambauer, G. and Hochreiter, S. and Nearing, G.},
 title = {Towards learning universal, regional, and local hydrological behaviors via machine learning applied to large-sample datasets},
 journal = {Hydrology and Earth System Sciences},
