### 2022-01-11 f611023
unify writing of author names
