+++ bibliography.bib
@@ -1,6 +1,6 @@
 @book{biswas1970history,
   title={History of hydrology},
-  author={Biswas, Asit K},
+  author={Biswas, A.},
   year={1970},
   publisher={North Holland Publishing Company},
   address={Amsterdam}
@@ -8,7 +8,7 @@
 
 @book{angelakis2012evolution,
   title={Evolution of Water Supply Through the Millennia},
-  author={Angelakis, Andreas and Mays, Larry and  Koutsoyiannis, Demetris and Mamassis, Nikos},
+  author={Angelakis, A. and Mays, L. and  Koutsoyiannis, D. and Mamassis, N.},
   year={2012},
   publisher={IWA Publishing},
   address={London}
@@ -35,7 +35,7 @@ doi = {10.5194/hess-23-5089-2019}
 }
 
 @article{kratzert2019pub,
-author = {Kratzert, Frederik and Klotz, Daniel and Herrnegger, Mathew and Sampson, Alden K. and Hochreiter, Sepp and Nearing, Grey S.},
+author = {Kratzert, F. and Klotz, D. and Herrnegger, M. and Sampson, A. and Hochreiter, S.and Nearing, G.},
 title = {Toward Improved Predictions in Ungauged Basins: Exploiting the Power of Machine Learning},
 journal = {Water Resources Research},
 volume = {55},
@@ -57,7 +57,7 @@ doi = {10.5194/hess-25-2045-2021}
 }
 
 @article{klotz2021uncertainty,
-author = {Klotz, D. and Kratzert, F. and Gauch, M. and Keefe Sampson, A. and Brandstetter, J. and Klambauer, G. and Hochreiter, S. and Nearing, G.},
+author = {Klotz, D. and Kratzert, F. and Gauch, M. and Sampson, A. and Brandstetter, J. and Klambauer, G. and Hochreiter, S. and Nearing, G.},
 title = {Uncertainty Estimation with Deep Learning for Rainfall–Runoff Modelling},
 journal = {Hydrology and Earth System Sciences Discussions},
 volume = {2021},
@@ -79,7 +79,7 @@ doi = {10.5194/hess-2021-127}
 }
 
 @article{nearing2021role,
-author = {Nearing, Grey S. and Kratzert, Frederik and Sampson, Alden Keefe and Pelissier, Craig S. and Klotz, Daniel and Frame, Jonathan M. and Prieto, Cristina and Gupta, Hoshin V.},
+author = {Nearing, G. and Kratzert, F. and Sampson, A. and Pelissier, C. and Klotz, D. and Frame, J. and Prieto, C. and Gupta, H.},
 title = {What Role Does Hydrological Science Play in the Age of Machine Learning?},
 journal = {Water Resources Research},
 volume = {57},
@@ -89,7 +89,7 @@ year = {2021}
 }
 
 @article{keith2020deep,
-author = {Beven, Keith},
+author = {Beven, K.},
 title = {Deep learning, hydrological processes and the uniqueness of place},
 journal = {Hydrological Processes},
 volume = {34},
