### 2022-01-11 #4050 @kratzert
Sorry for the spam but I just saw that we had different spellings for the same authors across different bib entries, which led to an undesired format in the reference list. I fixed this and will query whedon again to compile the manuscript. Please consider the document below for the review.
