+++ bibliography.bib
@@ -11,7 +11,8 @@
   author={Angelakis, A. and Mays, L. and  Koutsoyiannis, D. and Mamassis, N.},
   year={2012},
   publisher={IWA Publishing},
-  address={London}
+  address={London},
+  doi={10.2166/9781780401041},
 }
 
 @book{loague2010rainfall,
@@ -84,7 +85,7 @@ title = {What Role Does Hydrological Science Play in the Age of Machine Learning
 journal = {Water Resources Research},
 volume = {57},
 number = {3},
-doi = {https://doi.org/10.1029/2020WR028091},
+doi = {10.1029/2020WR028091},
 year = {2021}
 }
 
@@ -95,7 +96,7 @@ journal = {Hydrological Processes},
 volume = {34},
 number = {16},
 pages = {3608-3613},
-doi = {https://doi.org/10.1002/hyp.13805},
+doi = {10.1002/hyp.13805},
 year = {2020}
 }
 
