### 2022-01-25 #4050 @elbeejay
Hi all, just to check-in and clarify - that is a 2 week reminder above. As JOSS is still in a "[reduced service mode](https://blog.joss.theoj.org/2020/05/reopening-joss)" due to COVID, we are looking for your reviews to be complete within 6 weeks total, which would be 4 weeks from today (Feb. 22).

We are of course understanding of any potential delays or personal issues which may prevent you from conducting your review within that time-span, but ask that you please let us know ahead of time here if your review is going to be delayed.

Otherwise thank you again for reviewing for JOSS, and reach out to me if you have any questions as you conduct your reviews.
