### 2022-01-26 40c81a5
Fix git hash logging (#126)
With #114, the logging of the git hash broke, because the string decoding got lost.
To recover commit hashes that weren't logged correctly:
The encoded value will look something like `commit_hash: !!binary | ZTdiZTYwNQo=`. THe last part (`ZTdiZTYwNQo=`) is the base64-encoded commit hash that can be decoded with any online tool, e.g. https://www.base64decode.org/.
