### 2022-01-28 neuralhydrology/neuralhydrology#66 @jhamman
Publish releases on PyPI
The current documentation for this package directs users to install this package from source, i.e.:

```
pip install git+https://github.com/neuralhydrology/neuralhydrology.git
```

For ease of use and to better enable reproducible research, I think it would be valuable to also support installing neuralhydrology from PyPI and/or Conda-Forge.

xref: https://github.com/openjournals/joss-reviews/issues/4050
