### 2022-01-28 #4050 @jhamman
I'm grateful I was able to review this submission. I have a few comments relative to my review:

- The article indicates the package is installable via `pip` but upon inspection, the package has not been published on PyPI (or another package manager like conda-forge).  I've opened https://github.com/neuralhydrology/neuralhydrology/issues/66 to suggest this be done.
- There appears to be a missing citation under **Example Use Cases** in the article. See _Kratzert et al. 20xx_.
- The authors are encouraged to add a citation to the the following claim:
  > Users can easily train an ensemble of multiple models and average their predictions, which often greatly enhances the overall accuracy.
