### 2022-01-28 neuralhydrology/neuralhydrology#66 @kratzert
Hi @jhamman

I agree that we should support installation from PyPI (and Conda-Forge). I think we had this in mind initially but wanted to do this only when hitting the first major version. When we eventually reached that I had long forgotten about this. But great suggestion and I'll see what is required for this process and how to automate it.
