### 2022-01-28 neuralhydrology/neuralhydrology#66 @jhamman
👍 , FWIW, I find [Dask's release procedure](https://github.com/dask/dask/blob/main/docs/release-procedure.md) to be the right level of complexity for most projects. If you are looking for a simple way to get started, it's where I would start.
