### 2022-01-28 f562968
Add issue templates drafts (#124)
* Add issue templates drafts

* Add default title

* Update .github/ISSUE_TEMPLATE/bug_report.md

Co-authored-by: Martin Gauch <15731649+gauchm@users.noreply.github.com>

* Update bug_report.md

Co-authored-by: Martin Gauch <15731649+gauchm@users.noreply.github.com>
