+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,31 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: "[BUG] Title describing the bug"
+labels: ''
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Steps to reproduce the behavior. E.g. which data did you use, what were the commands that you executed, did you modify the code, etc.
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Description of the dataset**
+Provide details on the dataset that you use. Is it e.g. an out-of-the-box CAMELS dataset, or did you create your own csv/netCDF files. In that case, providing a data sample might be beneficial.
+
+**Logs & Screenshots**
+Please provide the full stack trace if any exception occured. If applicable, add screenshots to help explain your problem.
+
+**Desktop & Environment (please complete the following information):**
+ - OS: [e.g. Linux, Windows, iOS]
+ - The git commit if you cloned the repo, or the version number in `neuralhydrology/__about__.py`)
+ - The Python version and a list of installed Python packages. If you use conda, you can create this list via `conda env export`.
+
+**Additional context**
+Add any other context about the problem here.
