### 2022-02-05 neuralhydrology/neuralhydrology#68 @ammilten
Introduction python script difficult to follow
[Part of JOSS Review: https://github.com/openjournals/joss-reviews/issues/4050]

I'm not a deep learning expert and have only used very simple neural nets in my research. I would like to do more and this library seems very powerful. The authors have put an impressive amount of effort into this. However, I am having trouble getting started.

I'm trying to use the python API but I can't follow the [introduction tutorial](https://neuralhydrology.readthedocs.io/en/latest/tutorials/introduction.html). I get hung up on

`start_run(config_file=Path("1_basin.yml"))`

because there is no file "1_basin.yml". It would be very helpful to have default options in `start_run()` that automatically locates or creates a .yml file. If the problem is that the training data are stored on other servers, then maybe have a few shell commands that download the data would help. Or, the data could be trimmed enough that they could be stored in the git repo.

The main point is that making the introduction script easier for a new machine would make the entire library more accessible to the casual user.
