### 2022-02-05 neuralhydrology/neuralhydrology#68 @kratzert
Hi @ammilten

thanks for commenting and providing feedback, much appreciated.

Are you running the notebook locally on your computer? Because the `1_basin.yml` file is in the same folder as the notebook file (see [here](https://github.com/neuralhydrology/neuralhydrology/tree/master/examples/01-Introduction)). I just re-run the notebook on my local machine and had no problem with this section. Can you maybe paste the exact error that you are seeing?

I agree that our instructions for the [data requirements](https://neuralhydrology.readthedocs.io/en/latest/tutorials/introduction.html#Data-requirements) are not really helpful for people that are not familiar with the data set and we will fix that. So in case you don't have the data set, almost all of our tutorials require the publicly available [CAMELS](https://ral.ucar.edu/solutions/products/camels) data set and depending on where you store it, you need to adapt the `data_dir` argument in the [1_basin.yml](https://github.com/neuralhydrology/neuralhydrology/blob/master/examples/01-Introduction/1_basin.yml) file.

But again, I agree that this part is not very clear and we will push an update in the next days, thanks again for your feedback.
