### 2022-02-05 neuralhydrology/neuralhydrology#68 @ammilten
I installed the code using the following command from the install page.

`pip install git+https://github.com/neuralhydrology/neuralhydrology.git`

In retrospect I now see that there is a link to the notebook, which I simply missed. Usually when I am going through a new library I don't have time to read all the text and just skip straight to the code. Simply making the link clearer at the top of the page could make a big difference for potential users like me.

Now, I'm getting an error when running the script, again on the `start_run()` command. This time it's because my Torch isn't compiled with CUDA. How essential is this CUDA compatibility for the library beyond this example?

```
>>start_run(config_file=Path("1_basin.yml"))
2022-02-05 11:46:37,123: Logging to /Users/ammilten/JOSS Review/01-Introduction/runs/test_run_0502_114637/output.log initialized.
2022-02-05 11:46:37,125: ### Folder structure created at /Users/ammilten/JOSS Review/01-Introduction/runs/test_run_0502_114637
2022-02-05 11:46:37,125: ### Run configurations for test_run
2022-02-05 11:46:37,126: experiment_name: test_run
2022-02-05 11:46:37,127: train_basin_file: 1_basin.txt
2022-02-05 11:46:37,128: validation_basin_file: 1_basin.txt
2022-02-05 11:46:37,128: test_basin_file: 1_basin.txt
2022-02-05 11:46:37,129: train_start_date: 1999-10-01 00:00:00
2022-02-05 11:46:37,130: train_end_date: 2008-09-30 00:00:00
2022-02-05 11:46:37,131: validation_start_date: 1980-10-01 00:00:00
2022-02-05 11:46:37,131: validation_end_date: 1989-09-30 00:00:00
2022-02-05 11:46:37,132: test_start_date: 1989-10-01 00:00:00
2022-02-05 11:46:37,132: test_end_date: 1999-09-30 00:00:00
2022-02-05 11:46:37,133: device: cuda:0
2022-02-05 11:46:37,133: validate_every: 3
2022-02-05 11:46:37,134: validate_n_random_basins: 1
2022-02-05 11:46:37,135: metrics: ['NSE']
2022-02-05 11:46:37,135: model: cudalstm
2022-02-05 11:46:37,135: head: regression
2022-02-05 11:46:37,136: output_activation: linear
2022-02-05 11:46:37,136: hidden_size: 20
2022-02-05 11:46:37,137: initial_forget_bias: 3
2022-02-05 11:46:37,137: output_dropout: 0.4
2022-02-05 11:46:37,138: optimizer: Adam
2022-02-05 11:46:37,139: loss: MSE
2022-02-05 11:46:37,139: learning_rate: {0: 0.01, 30: 0.005, 40: 0.001}
2022-02-05 11:46:37,139: batch_size: 256
2022-02-05 11:46:37,140: epochs: 50
2022-02-05 11:46:37,140: clip_gradient_norm: 1
2022-02-05 11:46:37,141: predict_last_n: 1
2022-02-05 11:46:37,141: seq_length: 365
2022-02-05 11:46:37,142: num_workers: 8
2022-02-05 11:46:37,142: log_interval: 5
2022-02-05 11:46:37,143: log_tensorboard: True
2022-02-05 11:46:37,144: log_n_figures: 1
2022-02-05 11:46:37,144: save_weights_every: 1
2022-02-05 11:46:37,145: dataset: camels_us
2022-02-05 11:46:37,145: data_dir: ../../data/CAMELS_US
2022-02-05 11:46:37,145: forcings: ['maurer', 'daymet', 'nldas']
2022-02-05 11:46:37,146: dynamic_inputs: ['PRCP(mm/day)_nldas', 'PRCP(mm/day)_maurer', 'prcp(mm/day)_daymet', 'srad(W/m2)_daymet', 'tmax(C)_daymet', 'tmin(C)_daymet', 'vp(Pa)_daymet']
2022-02-05 11:46:37,147: target_variables: ['QObs(mm/d)']
2022-02-05 11:46:37,147: clip_targets_to_zero: ['QObs(mm/d)']
2022-02-05 11:46:37,148: number_of_basins: 1
2022-02-05 11:46:37,148: run_dir: /Users/ammilten/JOSS Review/01-Introduction/runs/test_run_0502_114637
2022-02-05 11:46:37,149: train_dir: /Users/ammilten/JOSS Review/01-Introduction/runs/test_run_0502_114637/train_data
2022-02-05 11:46:37,149: img_log_dir: /Users/ammilten/JOSS Review/01-Introduction/runs/test_run_0502_114637/img_log
2022-02-05 11:46:37,168: ### Device cuda:0 will be used for training
AssertionError                            Traceback (most recent call last)
<ipython-input-2-5c4ab3b99e82> in <module>
----> 1 start_run(config_file=Path("1_basin.yml"))

~/opt/anaconda3/lib/python3.8/site-packages/neuralhydrology/nh_run.py in start_run(config_file, gpu)
     74         config.device = "cpu"
     75
---> 76     start_training(config)
     77
     78

~/opt/anaconda3/lib/python3.8/site-packages/neuralhydrology/training/train.py in start_training(cfg)
     20     else:
     21         raise ValueError(f"Unknown head {cfg.head}.")
---> 22     trainer.initialize_training()
     23     trainer.train_and_validate()

~/opt/anaconda3/lib/python3.8/site-packages/neuralhydrology/training/basetrainer.py in initialize_training(self)
    135         If called in a ``continue_training`` context, this model will also restore the model and optimizer state.
    136         """
--> 137         self.model = self._get_model().to(self.device)
    138         if self.cfg.checkpoint_path is not None:
    139             LOGGER.info(f"Starting training from Checkpoint {self.cfg.checkpoint_path}")

~/opt/anaconda3/lib/python3.8/site-packages/torch/nn/modules/module.py in to(self, *args, **kwargs)
    897             return t.to(device, dtype if t.is_floating_point() or t.is_complex() else None, non_blocking)
    898
--> 899         return self._apply(convert)
    900
    901     def register_backward_hook(

~/opt/anaconda3/lib/python3.8/site-packages/torch/nn/modules/module.py in _apply(self, fn)
    568     def _apply(self, fn):
    569         for module in self.children():
--> 570             module._apply(fn)
    571
    572         def compute_should_use_set_data(tensor, tensor_applied):

~/opt/anaconda3/lib/python3.8/site-packages/torch/nn/modules/rnn.py in _apply(self, fn)
    180
    181     def _apply(self, fn):
--> 182         ret = super(RNNBase, self)._apply(fn)
    183
    184         # Resets _flat_weights

~/opt/anaconda3/lib/python3.8/site-packages/torch/nn/modules/module.py in _apply(self, fn)
    591             # `with torch.no_grad():`
    592             with torch.no_grad():
--> 593                 param_applied = fn(param)
    594             should_use_set_data = compute_should_use_set_data(param, param_applied)
    595             if should_use_set_data:

~/opt/anaconda3/lib/python3.8/site-packages/torch/nn/modules/module.py in convert(t)
    895                 return t.to(device, dtype if t.is_floating_point() or t.is_complex() else None,
    896                             non_blocking, memory_format=convert_to_format)
--> 897             return t.to(device, dtype if t.is_floating_point() or t.is_complex() else None, non_blocking)
    898
    899         return self._apply(convert)

~/opt/anaconda3/lib/python3.8/site-packages/torch/cuda/__init__.py in _lazy_init()
    206                 "multiprocessing, you must use the 'spawn' start method")
    207         if not hasattr(torch._C, '_cuda_getDeviceCount'):
--> 208             raise AssertionError("Torch not compiled with CUDA enabled")
    209         if _cudart is None:
    210             raise AssertionError(

AssertionError: Torch not compiled with CUDA enabled
```
