### 2022-02-05 neuralhydrology/neuralhydrology#68 @kratzert
@ammilten

again thanks for your feedback. We will make sure that we make the whole section about the setup/data set clearer.

Regarding CUDA: absolutely not required, especially for this tutorial. Just switch the [device](https://github.com/neuralhydrology/neuralhydrology/blob/31bd284c4713dce1404818f9416161303c25b436/examples/01-Introduction/1_basin.yml#L20) argument in the `1_basin.yml` config to `cpu`. Again, this should also probably be clearer, so thanks again for your feedback.
