### 2022-02-05 neuralhydrology/neuralhydrology#68 @ammilten
Thank you @kratzert. I have one last comment, and then I'll pause to give you a chance to make some changes to the setup/intro script documentation. At first glance the rest of the code appears to have good documentation and functionality. If we can get to a point where the intro script is easy to execute, then we should be in great shape.

I got a little further on the intro script but having a new issue. I changed the device argument to `cpu` and downloaded the CAMELS dataset from [here](https://ral.ucar.edu/solutions/products/camels). Now I'm getting a `FileNotFoundError: No file for Basin 01022500 at []`.

Here's my directory structure, and the modified .yml file is attached. Do you think this has to do with how I downloaded the data?
```
01-Introduction
-- 1_basin.txt
-- 1_basin.yml
-- basin_mean_forcing //extracted from "CAMELS time series meteorology, observed flow, meta data (.zip)"
-- Introduction.ipynb
```

Full stacktrace:
```
>> start_run(config_file=Path("1_basin.yml"))
2022-02-05 12:13:15,933: Logging to /Users/ammilten/JOSS Review/01-Introduction/runs/test_run_0502_121315/output.log initialized.
2022-02-05 12:13:15,934: ### Folder structure created at /Users/ammilten/JOSS Review/01-Introduction/runs/test_run_0502_121315
2022-02-05 12:13:15,935: ### Run configurations for test_run
2022-02-05 12:13:15,935: experiment_name: test_run
2022-02-05 12:13:15,936: train_basin_file: 1_basin.txt
2022-02-05 12:13:15,936: validation_basin_file: 1_basin.txt
2022-02-05 12:13:15,937: test_basin_file: 1_basin.txt
2022-02-05 12:13:15,937: train_start_date: 1999-10-01 00:00:00
2022-02-05 12:13:15,938: train_end_date: 2008-09-30 00:00:00
2022-02-05 12:13:15,939: validation_start_date: 1980-10-01 00:00:00
2022-02-05 12:13:15,939: validation_end_date: 1989-09-30 00:00:00
2022-02-05 12:13:15,940: test_start_date: 1989-10-01 00:00:00
2022-02-05 12:13:15,941: test_end_date: 1999-09-30 00:00:00
2022-02-05 12:13:15,941: device: cpu
2022-02-05 12:13:15,942: validate_every: 3
2022-02-05 12:13:15,942: validate_n_random_basins: 1
2022-02-05 12:13:15,943: metrics: ['NSE']
2022-02-05 12:13:15,943: model: cudalstm
2022-02-05 12:13:15,944: head: regression
2022-02-05 12:13:15,944: output_activation: linear
2022-02-05 12:13:15,945: hidden_size: 20
2022-02-05 12:13:15,945: initial_forget_bias: 3
2022-02-05 12:13:15,946: output_dropout: 0.4
2022-02-05 12:13:15,947: optimizer: Adam
2022-02-05 12:13:15,947: loss: MSE
2022-02-05 12:13:15,948: learning_rate: {0: 0.01, 30: 0.005, 40: 0.001}
2022-02-05 12:13:15,949: batch_size: 256
2022-02-05 12:13:15,950: epochs: 50
2022-02-05 12:13:15,951: clip_gradient_norm: 1
2022-02-05 12:13:15,951: predict_last_n: 1
2022-02-05 12:13:15,952: seq_length: 365
2022-02-05 12:13:15,953: num_workers: 8
2022-02-05 12:13:15,954: log_interval: 5
2022-02-05 12:13:15,954: log_tensorboard: True
2022-02-05 12:13:15,955: log_n_figures: 1
2022-02-05 12:13:15,957: save_weights_every: 1
2022-02-05 12:13:15,957: dataset: camels_us
2022-02-05 12:13:15,958: data_dir: .
2022-02-05 12:13:15,958: forcings: ['maurer', 'daymet', 'nldas']
2022-02-05 12:13:15,959: dynamic_inputs: ['PRCP(mm/day)_nldas', 'PRCP(mm/day)_maurer', 'prcp(mm/day)_daymet', 'srad(W/m2)_daymet', 'tmax(C)_daymet', 'tmin(C)_daymet', 'vp(Pa)_daymet']
2022-02-05 12:13:15,960: target_variables: ['QObs(mm/d)']
2022-02-05 12:13:15,960: clip_targets_to_zero: ['QObs(mm/d)']
2022-02-05 12:13:15,961: number_of_basins: 1
2022-02-05 12:13:15,961: run_dir: /Users/ammilten/JOSS Review/01-Introduction/runs/test_run_0502_121315
2022-02-05 12:13:15,962: train_dir: /Users/ammilten/JOSS Review/01-Introduction/runs/test_run_0502_121315/train_data
2022-02-05 12:13:15,963: img_log_dir: /Users/ammilten/JOSS Review/01-Introduction/runs/test_run_0502_121315/img_log
2022-02-05 12:13:15,978: ### Device cpu will be used for training
2022-02-05 12:13:15,981: Loading basin data into xarray data set.
  0%|          | 0/1 [00:00<?, ?it/s]2022-02-05 12:13:16,179: Note: NumExpr detected 16 cores but "NUMEXPR_MAX_THREADS" not set, so enforcing safe limit of 8.
2022-02-05 12:13:16,179: NumExpr defaulting to 8 threads.
  0%|          | 0/1 [00:00<?, ?it/s]
FileNotFoundError                         Traceback (most recent call last)
<ipython-input-5-5c4ab3b99e82> in <module>
----> 1 start_run(config_file=Path("1_basin.yml"))

~/opt/anaconda3/lib/python3.8/site-packages/neuralhydrology/nh_run.py in start_run(config_file, gpu)
     74         config.device = "cpu"
     75
---> 76     start_training(config)
     77
     78

~/opt/anaconda3/lib/python3.8/site-packages/neuralhydrology/training/train.py in start_training(cfg)
     20     else:
     21         raise ValueError(f"Unknown head {cfg.head}.")
---> 22     trainer.initialize_training()
     23     trainer.train_and_validate()

~/opt/anaconda3/lib/python3.8/site-packages/neuralhydrology/training/basetrainer.py in initialize_training(self)
    160             self._restore_training_state()
    161
--> 162         ds = self._get_dataset()
    163         if len(ds) == 0:
    164             raise ValueError("Dataset contains no samples.")

~/opt/anaconda3/lib/python3.8/site-packages/neuralhydrology/training/basetrainer.py in _get_dataset(self)
     77
     78     def _get_dataset(self) -> BaseDataset:
---> 79         return get_dataset(cfg=self.cfg, period="train", is_train=True, scaler=self._scaler)
     80
     81     def _get_model(self) -> torch.nn.Module:

~/opt/anaconda3/lib/python3.8/site-packages/neuralhydrology/datasetzoo/__init__.py in get_dataset(cfg, is_train, period, basin, additional_features, id_to_int, scaler)
     67         raise NotImplementedError(f"No dataset class implemented for dataset {cfg.dataset}")
     68
---> 69     ds = Dataset(cfg=cfg,
     70                  is_train=is_train,
     71                  period=period,

~/opt/anaconda3/lib/python3.8/site-packages/neuralhydrology/datasetzoo/camelsus.py in __init__(self, cfg, is_train, period, basin, additional_features, id_to_int, scaler)
     58                  id_to_int: Dict[str, int] = {},
     59                  scaler: Dict[str, Union[pd.Series, xarray.DataArray]] = {}):
---> 60         super(CamelsUS, self).__init__(cfg=cfg,
     61                                        is_train=is_train,
     62                                        period=period,

~/opt/anaconda3/lib/python3.8/site-packages/neuralhydrology/datasetzoo/basedataset.py in __init__(self, cfg, is_train, period, basin, additional_features, id_to_int, scaler)
    129
    130         # load and preprocess data
--> 131         self._load_data()
    132
    133         if self.is_train:

~/opt/anaconda3/lib/python3.8/site-packages/neuralhydrology/datasetzoo/basedataset.py in _load_data(self)
    558         self._load_combined_attributes()
    559
--> 560         xr = self._load_or_create_xarray_dataset()
    561
    562         if self.cfg.loss.lower() in ['nse', 'weightednse']:

~/opt/anaconda3/lib/python3.8/site-packages/neuralhydrology/datasetzoo/basedataset.py in _load_or_create_xarray_dataset(self)
    273                 LOGGER.info("Loading basin data into xarray data set.")
    274             for basin in tqdm(self.basins, disable=self._disable_pbar, file=sys.stdout):
--> 275                 df = self._load_basin_data(basin)
    276
    277                 # add columns from dataframes passed as additional data files

~/opt/anaconda3/lib/python3.8/site-packages/neuralhydrology/datasetzoo/camelsus.py in _load_basin_data(self, basin)
     80
     81         # add discharge
---> 82         df['QObs(mm/d)'] = load_camels_us_discharge(self.cfg.data_dir, basin, area)
     83
     84         # replace invalid discharge values by NaNs

~/opt/anaconda3/lib/python3.8/site-packages/neuralhydrology/datasetzoo/camelsus.py in load_camels_us_discharge(data_dir, basin, area)
    215         file_path = file_path[0]
    216     else:
--> 217         raise FileNotFoundError(f'No file for Basin {basin} at {file_path}')
    218
    219     col_names = ['basin', 'Year', 'Mnth', 'Day', 'QObs', 'flag']

FileNotFoundError: No file for Basin 01022500 at []
```

[1_basin.yml.zip](https://github.com/neuralhydrology/neuralhydrology/files/8008845/1_basin.yml.zip)
