### 2022-02-05 neuralhydrology/neuralhydrology#68 @kratzert
What you are missing is the streamflow folder (the target data). You only extracted the meteorological time series data, which is the model input, but not the streamflow folder (`usgs_streamflow`).

In general, the [CAMELS US data set class](https://github.com/neuralhydrology/neuralhydrology/blob/master/neuralhydrology/datasetzoo/camelsus.py) looks for the the following folder structure:

```
CAMELS_US/
- basin_mean_forcing/  # contains various time series of meteorological time series
- usgs_streamflow/  # contains streamflow time series
- camels_attributes_v2.0/  # downloaded from the same page at the bottom, used in later tutorials.
```

Store them anywhere and make sure the `data_dir` argument in the config points to the root directory that contains these three subdirectories.
