### 2022-02-09 neuralhydrology/neuralhydrology#68 @kratzert
Hi @ammilten

We pushed an update of the tutorials to this public repository and I hope it will make things easier. If you have time, it  would be awesome to get your feedback.

Specifically, we
- added a new tutorial on [Data Prerequisites](https://neuralhydrology.readthedocs.io/en/latest/tutorials/data-prerequisites.html) that explains in all details which data to download and how to organize the CAMELS US dataset locally
- Linked this data tutorial in all the other tutorials at the very top and tried to make the link to the source files more obvious. (see e.g. the very top of the [Introduction tutorial](https://neuralhydrology.readthedocs.io/en/latest/tutorials/introduction.html)
- Change a few lines of code so that the notebooks run on CPU-only environments without any changes to the config files (and added a text explanation). See e.g. [this section](https://neuralhydrology.readthedocs.io/en/latest/tutorials/introduction.html#Train-a-model-for-a-single-config-file) (similar in all tutorials)
