### 2022-02-12 22684ec
Minor bugfixes in transformer, typos (#134)
Closes #125
