### 2022-02-12 dd37e67
Add missing config arg allow_subsequent_nan_losses, closes #122 (#133)
* Add missing config arg allow_subsequent_nan_losses, closes #122

* Update docs/source/usage/config.rst

Co-authored-by: Martin Gauch <15731649+gauchm@users.noreply.github.com>

Co-authored-by: Martin Gauch <15731649+gauchm@users.noreply.github.com>
