### 2022-02-13 #4050 @ammilten
I really enjoyed going through this code. Once I got the intro script working and figured out the structure of the code, I really started to appreciate the amount of effort the author put in to making deep learning accessible. The way the code uses .yml files to set the training, etc. settings makes it really easy to work with deep learning.

Before publishing, the only thing I see that needs to be addressed is the missing citation in the paper (e.g. Kratzert et al. 20xx)

Once this is addressed I support accepting this submission.

A few suggested features for the future:
1. Python wrappers for reading/writing .yml files would be very useful.
2. Automated tests, if possible
3. Model analysis: Explainability (e.g. LIME/SHAP), feature importances, uncertainty quantification (e.g. Monte Carlo dropout). These would be lots of work but if you ever use them in your own research it would be cool to see them added to this library.

These are just suggestions and not necessary before acceptance. Congratulations to the authors for a high-quality, comprehensive piece of work.

Alex
