### 2022-02-13 93dfa02
Update installation guide for PyPI (#135)
follow-up on #131
