+++ paper.md
@@ -53,7 +53,7 @@ Lastly, the modular design of the library allows contributors not only to extend
 
 ## Example use-cases
 
-The most basic use-case for the NeuralHydrology library is basic rainfall--runoff modeling as described in Kratzert et al. 20xx, where we train a machine learning model on meteorological data and discharge observations from a set of basins (a process also known as calibration) and subsequently apply the model to a different time period, either to evaluate its goodness-of-fit, or to actually generate predictions for practical use. Since the experiment configuration is highly flexible, users can define which inputs they want to ingest into the model (e.g., whether discharge from previous time steps should be an input variable) and which variable(s) they want to predict. Users can easily train an ensemble of multiple models and average their predictions, which often greatly enhances the overall accuracy.
+The most basic use-case for the NeuralHydrology library is basic rainfall--runoff modeling as described in @kratzert2018rainfall and @kratzert2019regional, where we train a machine learning model on meteorological data and discharge observations from a set of basins (a process also known as calibration) and subsequently apply the model to a different time period, either to evaluate its goodness-of-fit, or to actually generate predictions for practical use. Since the experiment configuration is highly flexible, users can define which inputs they want to ingest into the model (e.g., whether discharge from previous time steps should be an input variable) and which variable(s) they want to predict. Users can easily train an ensemble of multiple models and average their predictions, which often greatly enhances the overall accuracy (see e.g. @kratzert2019regional in the context of LSTM-based rainfall-runoff modeling).
 
 Beyond this most basic application, NeuralHydrology also supports prediction in ungauged basins, where a model is trained on one set of basins and subsequently applied to another set of basins for which no discharge observations exist (see e.g. @kratzert2019pub). Lastly, NeuralHydrology also includes different options to train neural networks for uncertainty estimation, where the model does not only regress a single target value (e.g. discharge) per time step, but rather outputs probability distributions for each target variable (for details, see @klotz2021uncertainty).
 
