+++ bibliography.bib
@@ -111,4 +111,15 @@ year = {2020}
  year = {2019}
 }
 
+@Article{kratzert2018rainfall,
+author = {Kratzert, F. and Klotz, D. and Brenner, C. and Schulz, K. and Herrnegger, M.},
+title = {Rainfall--runoff modelling using Long Short-Term Memory (LSTM) networks},
+journal = {Hydrology and Earth System Sciences},
+volume = {22},
+year = {2018},
+number = {11},
+pages = {6005--6022},
+doi = {10.5194/hess-22-6005-2018}
+}
+
 
