### 2022-02-13 #4050 @kratzert
> I really enjoyed going through this code. Once I got the intro script working and figured out the structure of the code, I really started to appreciate the amount of effort the author put in to making deep learning accessible. The way the code uses .yml files to set the training, etc. settings makes it really easy to work with deep learning.
>
> Before publishing, the only thing I see that needs to be addressed is the missing citation in the paper (e.g. Kratzert et al. 20xx)
>
> Once this is addressed I support accepting this submission.
>
> A few suggested features for the future:
>
>     1. Python wrappers for reading/writing .yml files would be very useful.
>
>     2. Automated tests, if possible
>
>     3. Model analysis: Explainability (e.g. LIME/SHAP), feature importances, uncertainty quantification (e.g. Monte Carlo dropout). These would be lots of work but if you ever use them in your own research it would be cool to see them added to this library.
>
>
> These are just suggestions and not necessary before acceptance. Congratulations to the authors for a high-quality, comprehensive piece of work.
>
> Alex

Thanks @ammilten for your time and your valuable feedback on our tutorials.

- The reference is fixed, see above (and I will invoke the bot to recompile the paper for this thread).

Regarding your suggestions:
- Not sure if this is what you mean, but we already have utilities for reading/writing .yml files. Most of it is related to our `Config` class (see [docs](https://neuralhydrology.readthedocs.io/en/latest/api/neuralhydrology.utils.config.html)), which we use everywhere in the code to parse the .yml config arguments. The class has a public API for reading yaml files (the object is initialized from the yaml), to [change/update](https://neuralhydrology.readthedocs.io/en/latest/api/neuralhydrology.utils.config.html#neuralhydrology.utils.config.Config.update_config) the config arguments (via the `.update_config()` method, and to [store](https://neuralhydrology.readthedocs.io/en/latest/api/neuralhydrology.utils.config.html#neuralhydrology.utils.config.Config.dump_config) the config to a yaml file (via the `dump_config()` method, which deals with translating Path objects to string and converting pandas Timestamps to strings). Furthermore, there is also a `as_dict()` method, in case anyone wants to work with the parsed yaml config as a Python dictionary. We also have the [create_config_files](https://neuralhydrology.readthedocs.io/en/latest/api/neuralhydrology.utils.configutils.html#neuralhydrology.utils.configutils.create_config_files) function that can be used to automatically create yaml files (given a base config) for e.g. hyper parameter tuning. Is there anything else you would like to see? Very happy for feedback!
- I agree on the tests. We have a couple of tests and they also run on every PR/push automatically via Github actions but we definitely should extend our test coverage (it is a TODO on our list...)
- Good suggestions, thanks. We did some related studies in the past (on explanability) but never integrated the specific functions into the library (but instead published them in a separate repository with the paper code). I agree that it makes totally sense to include such utilities also into the library. When we started, our vision for NeuralHydrology was more about doing quick experiments with different model architectures/datasets/settings and the entire part of the post-training analysis was always separated. Regarding uncertainty: This is actually something that is well integrated into the library already and we also published a paper about it. We have a couple of different options that one can use: Different mixture density heads ([GMM](https://neuralhydrology.readthedocs.io/en/latest/usage/models.html#gmm), [CMAL](https://neuralhydrology.readthedocs.io/en/latest/usage/models.html#cmal), [UMAL](https://neuralhydrology.readthedocs.io/en/latest/usage/models.html#umal)) that all return probabilistic outputs (instead of a point estimate in the standard regression setting) and we also support MC-Dropout sampling during evaluation (see `mc_dropout` under the [in the list of config arguments](https://neuralhydrology.readthedocs.io/en/latest/usage/config.html)). We also plan to release a tutorial on all of these options, as soon as the [corresponding paper](https://hess.copernicus.org/preprints/hess-2021-154/) is accepted for publication.

Again, thank you so much for your time reviewing our library/paper and for your feedback so far!
