### 2022-02-18 neuralhydrology/neuralhydrology#75 @chuckaustin
Trouble running MTS notebook
[Part of JOSS Review: https://github.com/https://github.com/openjournals/joss-reviews/issues/4050]

I've been going through the examples, and so far haven't run into much trouble. However, I'm getting stuck on Example 4 ("Multi-Timescale Prediction"). When I try to train the model using either nh-run or the python API, the process aborts right before the first epoch.

Copying output below. Have I missed something?

```
(neuralhydrology) C:\Users\321319\Documents\neuralhydrology\examples\04-Multi-Timescale>nh-run train --config-file 1_basin.yml
2022-02-17 17:27:05,401: Logging to C:\Users\321319\Documents\neuralhydrology\examples\04-Multi-Timescale\runs\test_run_1702_172705\output.log initialized.
2022-02-17 17:27:05,401: ### Folder structure created at C:\Users\321319\Documents\neuralhydrology\examples\04-Multi-Timescale\runs\test_run_1702_172705
2022-02-17 17:27:05,412: ### Run configurations for test_run
2022-02-17 17:27:05,432: experiment_name: test_run
2022-02-17 17:27:05,465: use_frequencies: ['1H', '1D']
2022-02-17 17:27:05,501: train_basin_file: 1_basin.txt
2022-02-17 17:27:05,516: validation_basin_file: 1_basin.txt
2022-02-17 17:27:05,548: test_basin_file: 1_basin.txt
2022-02-17 17:27:05,560: train_start_date: 1999-10-01 00:00:00
2022-02-17 17:27:05,585: train_end_date: 2008-09-30 00:00:00
2022-02-17 17:27:05,618: validation_start_date: 1996-10-01 00:00:00
2022-02-17 17:27:05,653: validation_end_date: 1999-09-30 00:00:00
2022-02-17 17:27:05,696: test_start_date: 1989-10-01 00:00:00
2022-02-17 17:27:05,733: test_end_date: 1996-09-30 00:00:00
2022-02-17 17:27:05,742: device: cpu
2022-02-17 17:27:05,769: validate_every: 5
2022-02-17 17:27:05,780: validate_n_random_basins: 1
2022-02-17 17:27:05,808: metrics: ['NSE']
2022-02-17 17:27:05,820: model: mtslstm
2022-02-17 17:27:05,849: shared_mtslstm: False
2022-02-17 17:27:05,882: transfer_mtslstm_states: {'h': 'linear', 'c': 'linear'}
2022-02-17 17:27:05,882: head: regression
2022-02-17 17:27:05,901: output_activation: linear
2022-02-17 17:27:05,915: hidden_size: 20
2022-02-17 17:27:05,923: initial_forget_bias: 3
2022-02-17 17:27:05,946: output_dropout: 0.4
2022-02-17 17:27:05,980: optimizer: Adam
2022-02-17 17:27:06,013: loss: MSE
2022-02-17 17:27:06,021: regularization: ['tie_frequencies']
2022-02-17 17:27:06,046: learning_rate: {0: 0.01, 30: 0.005, 40: 0.001}
2022-02-17 17:27:06,079: batch_size: 256
2022-02-17 17:27:06,112: epochs: 50
2022-02-17 17:27:06,124: clip_gradient_norm: 1
2022-02-17 17:27:06,146: predict_last_n: {'1D': 1, '1H': 24}
2022-02-17 17:27:06,148: seq_length: {'1D': 365, '1H': 336}
2022-02-17 17:27:06,162: num_workers: 8
2022-02-17 17:27:06,180: log_interval: 5
2022-02-17 17:27:06,187: log_tensorboard: False
2022-02-17 17:27:06,213: log_n_figures: 0
2022-02-17 17:27:06,246: save_weights_every: 1
2022-02-17 17:27:06,253: dataset: hourly_camels_us
2022-02-17 17:27:06,279: data_dir: ..\..\data\CAMELS_US
2022-02-17 17:27:06,313: forcings: ['nldas_hourly', 'daymet']
2022-02-17 17:27:06,346: dynamic_inputs: {'1D': ['prcp(mm/day)_daymet', 'srad(W/m2)_daymet', 'tmax(C)_daymet', 'tmin(C)_daymet', 'vp(Pa)_daymet'], '1H': ['convective_fraction_nldas_hourly', 'longwave_radiation_nldas_hourly', 'potential_energy_nldas_hourly', 'potential_evaporation_nldas_hourly', 'pressure_nldas_hourly', 'shortwave_radiation_nldas_hourly', 'specific_humidity_nldas_hourly', 'temperature_nldas_hourly', 'total_precipitation_nldas_hourly', 'wind_u_nldas_hourly', 'wind_v_nldas_hourly', 'prcp(mm/day)_daymet', 'srad(W/m2)_daymet', 'tmax(C)_daymet', 'tmin(C)_daymet', 'vp(Pa)_daymet']}
2022-02-17 17:27:06,381: target_variables: ['qobs_mm_per_hour']
2022-02-17 17:27:06,413: clip_targets_to_zero: ['qobs_mm_per_hour']
2022-02-17 17:27:06,423: number_of_basins: 1
2022-02-17 17:27:06,446: run_dir: C:\Users\321319\Documents\neuralhydrology\examples\04-Multi-Timescale\runs\test_run_1702_172705
2022-02-17 17:27:06,459: train_dir: C:\Users\321319\Documents\neuralhydrology\examples\04-Multi-Timescale\runs\test_run_1702_172705\train_data
2022-02-17 17:27:06,480: img_log_dir: C:\Users\321319\Documents\neuralhydrology\examples\04-Multi-Timescale\runs\test_run_1702_172705\img_log
2022-02-17 17:27:06,518: ### Device cpu will be used for training
2022-02-17 17:27:06,525: No specific hidden size for frequencies are specified. Same hidden size is used for all.
2022-02-17 17:27:06,553: Loading basin data into xarray data set.
100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 1/1 [00:00<00:00,  1.46it/s]
2022-02-17 17:27:07,312: Create lookup table and convert to pytorch tensor
100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 1/1 [00:01<00:00,  1.61s/it]
# Epoch 1:   0%|                                                                                                                                                               | 0/11 [00:00<?, ?it/s]
2022-02-17 17:27:09,078: Uncaught exception
Traceback (most recent call last):
  File "C:\Users\321319\Miniconda3\envs\neuralhydrology\Scripts\nh-run-script.py", line 33, in <module>
    sys.exit(load_entry_point('neuralhydrology', 'console_scripts', 'nh-run')())
  File "c:\users\321319\documents\neuralhydrology\neuralhydrology\nh_run.py", line 43, in _main
    start_run(config_file=Path(args["config_file"]), gpu=args["gpu"])
  File "c:\users\321319\documents\neuralhydrology\neuralhydrology\nh_run.py", line 76, in start_run
    start_training(config)
  File "c:\users\321319\documents\neuralhydrology\neuralhydrology\training\train.py", line 23, in start_training
    trainer.train_and_validate()
  File "c:\users\321319\documents\neuralhydrology\neuralhydrology\training\basetrainer.py", line 205, in train_and_validate
    self._train_epoch(epoch=epoch)
  File "c:\users\321319\documents\neuralhydrology\neuralhydrology\training\basetrainer.py", line 272, in _train_epoch
    for data in pbar:
  File "C:\Users\321319\Miniconda3\envs\neuralhydrology\lib\site-packages\tqdm\std.py", line 1180, in __iter__
    for obj in iterable:
  File "C:\Users\321319\Miniconda3\envs\neuralhydrology\lib\site-packages\torch\utils\data\dataloader.py", line 359, in __iter__
    return self._get_iterator()
  File "C:\Users\321319\Miniconda3\envs\neuralhydrology\lib\site-packages\torch\utils\data\dataloader.py", line 305, in _get_iterator
    return _MultiProcessingDataLoaderIter(self)
  File "C:\Users\321319\Miniconda3\envs\neuralhydrology\lib\site-packages\torch\utils\data\dataloader.py", line 918, in __init__
    w.start()
  File "C:\Users\321319\Miniconda3\envs\neuralhydrology\lib\multiprocessing\process.py", line 112, in start
    self._popen = self._Popen(self)
  File "C:\Users\321319\Miniconda3\envs\neuralhydrology\lib\multiprocessing\context.py", line 223, in _Popen
    return _default_context.get_context().Process._Popen(process_obj)
  File "C:\Users\321319\Miniconda3\envs\neuralhydrology\lib\multiprocessing\context.py", line 322, in _Popen
    return Popen(process_obj)
  File "C:\Users\321319\Miniconda3\envs\neuralhydrology\lib\multiprocessing\popen_spawn_win32.py", line 89, in __init__
    reduction.dump(process_obj, to_child)
  File "C:\Users\321319\Miniconda3\envs\neuralhydrology\lib\multiprocessing\reduction.py", line 60, in dump
    ForkingPickler(file, protocol).dump(obj)
TypeError: can't pickle _thread.lock objects
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "C:\Users\321319\Miniconda3\envs\neuralhydrology\lib\multiprocessing\spawn.py", line 105, in spawn_main
    exitcode = _main(fd)
  File "C:\Users\321319\Miniconda3\envs\neuralhydrology\lib\multiprocessing\spawn.py", line 115, in _main
    self = reduction.pickle.load(from_parent)
EOFError: Ran out of input
```
