### 2022-02-18 neuralhydrology/neuralhydrology#76 @chuckaustin
Trouble getting code to access GPU
[Part of JOSS Review: https://github.com/https://github.com/https://github.com/openjournals/joss-reviews/issues/4050]

One other issue is that I noticed that I'm unable to use my GPU (Nvidia Quadro P520) with the library.

I created an environment for neural hydrology on conda using environment_cuda9_2.yml. But when I try to train a model on my GPU I get an error saying pytorch was compiled without CUDA enabled.

```
2022-02-17 17:44:47,097: Uncaught exception
Traceback (most recent call last):
  File "C:\Users\321319\Miniconda3\envs\neuralhydrology\Scripts\nh-run-script.py", line 33, in <module>
    sys.exit(load_entry_point('neuralhydrology', 'console_scripts', 'nh-run')())
  File "c:\users\321319\documents\neuralhydrology\neuralhydrology\nh_run.py", line 43, in _main
    start_run(config_file=Path(args["config_file"]), gpu=args["gpu"])
  File "c:\users\321319\documents\neuralhydrology\neuralhydrology\nh_run.py", line 76, in start_run
    start_training(config)
  File "c:\users\321319\documents\neuralhydrology\neuralhydrology\training\train.py", line 22, in start_training
    trainer.initialize_training()
  File "c:\users\321319\documents\neuralhydrology\neuralhydrology\training\basetrainer.py", line 137, in initialize_training
    self.model = self._get_model().to(self.device)
  File "C:\Users\321319\Miniconda3\envs\neuralhydrology\lib\site-packages\torch\nn\modules\module.py", line 899, in to
    return self._apply(convert)
  File "C:\Users\321319\Miniconda3\envs\neuralhydrology\lib\site-packages\torch\nn\modules\module.py", line 570, in _apply
    module._apply(fn)
  File "C:\Users\321319\Miniconda3\envs\neuralhydrology\lib\site-packages\torch\nn\modules\module.py", line 570, in _apply
    module._apply(fn)
  File "C:\Users\321319\Miniconda3\envs\neuralhydrology\lib\site-packages\torch\nn\modules\rnn.py", line 182, in _apply
    ret = super(RNNBase, self)._apply(fn)
  File "C:\Users\321319\Miniconda3\envs\neuralhydrology\lib\site-packages\torch\nn\modules\module.py", line 593, in _apply
    param_applied = fn(param)
  File "C:\Users\321319\Miniconda3\envs\neuralhydrology\lib\site-packages\torch\nn\modules\module.py", line 897, in convert
    return t.to(device, dtype if t.is_floating_point() or t.is_complex() else None, non_blocking)
  File "C:\Users\321319\Miniconda3\envs\neuralhydrology\lib\site-packages\torch\cuda\__init__.py", line 208, in _lazy_init
    raise AssertionError("Torch not compiled with CUDA enabled")
AssertionError: Torch not compiled with CUDA enabled
```
