### 2022-02-23 neuralhydrology/neuralhydrology#75 @chuckaustin
Hi @gauchm, thanks for your response. I've made it through all the examples now except for the last one (which I have started, but haven't finished yet, as I'm running it on a CPU). The MTS notebook is the only place I get this particular error.

Setting num_workers to 0 made it execute! It also makes most of the other examples run faster on my machine, which I guess is because I'm running it on a single CPU.

Here is the output from conda env export. Would be awesome if there's an easy fix to make the software run on my GPU. As I mentioned in my other post, I created the environment using environment_cuda9_2.yml

```
(neuralhydrology) C:\Users\321319\Documents\neuralhydrology\examples\06-Finetuning>conda env export
name: neuralhydrology
channels:
  - pytorch
  - anaconda
  - defaults
dependencies:
  - alabaster=0.7.12=py37_0
  - argcomplete=1.12.3=pyhd3eb1b0_0
  - argon2-cffi=21.3.0=pyhd3eb1b0_0
  - argon2-cffi-bindings=21.2.0=py37h2bbff1b_0
  - async_generator=1.10=py37h28b3542_0
  - atomicwrites=1.4.0=py_0
  - attrs=21.4.0=pyhd3eb1b0_0
  - babel=2.9.1=pyhd3eb1b0_0
  - backcall=0.2.0=pyhd3eb1b0_0
  - blas=1.0=mkl
  - bleach=4.1.0=pyhd3eb1b0_0
  - bokeh=2.4.2=py37haa95532_0
  - bottleneck=1.3.2=py37h2a96729_1
  - brotli=1.0.9=ha925a31_2
  - brotlipy=0.7.0=py37h2bbff1b_1003
  - ca-certificates=2021.10.26=haa95532_4
  - cached-property=1.5.2=py_0
  - certifi=2021.10.8=py37haa95532_2
  - cffi=1.15.0=py37h2bbff1b_1
  - cftime=1.5.1.1=py37h080aedc_0
  - charset-normalizer=2.0.4=pyhd3eb1b0_0
  - colorama=0.4.4=pyhd3eb1b0_0
  - coverage=5.5=py37h2bbff1b_2
  - cryptography=36.0.0=py37h21b164f_0
  - cudatoolkit=9.2=0
  - curl=7.55.1=vc14hdaba4a4_3
  - cycler=0.11.0=pyhd3eb1b0_0
  - debugpy=1.5.1=py37hd77b12b_0
  - decorator=5.1.1=pyhd3eb1b0_0
  - defusedxml=0.7.1=pyhd3eb1b0_0
  - docutils=0.17.1=py37haa95532_1
  - entrypoints=0.3=py37_0
  - fonttools=4.25.0=pyhd3eb1b0_0
  - freetype=2.10.4=hd328e21_0
  - h5py=3.6.0=py37h3de5c98_0
  - hdf4=4.2.13=h712560f_2
  - hdf5=1.10.6=h7ebc959_0
  - icc_rt=2019.0.0=h0cc432a_1
  - icu=58.2=vc14hc45fdbb_0
  - idna=3.3=pyhd3eb1b0_0
  - imagesize=1.3.0=pyhd3eb1b0_0
  - importlib-metadata=4.8.2=py37haa95532_0
  - importlib_metadata=4.8.2=hd3eb1b0_0
  - iniconfig=1.1.1=pyhd3eb1b0_0
  - intel-openmp=2021.4.0=haa95532_3556
  - ipykernel=6.4.1=py37haa95532_1
  - ipython=7.31.1=py37haa95532_0
  - ipython_genutils=0.2.0=pyhd3eb1b0_1
  - ipywidgets=7.6.5=pyhd3eb1b0_1
  - jedi=0.18.1=py37haa95532_1
  - jinja2=3.0.2=pyhd3eb1b0_0
  - jpeg=9b=vc14h4d7706e_1
  - jsonschema=3.2.0=pyhd3eb1b0_2
  - jupyter=1.0.0=py37_7
  - jupyter_client=7.1.2=pyhd3eb1b0_0
  - jupyter_console=6.4.0=pyhd3eb1b0_0
  - jupyter_core=4.9.1=py37haa95532_0
  - jupyterlab_pygments=0.1.2=py_0
  - jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
  - kiwisolver=1.3.2=py37hd77b12b_0
  - libcurl=7.80.0=h86230a5_0
  - libnetcdf=4.6.1=hf59b723_4
  - libpng=1.6.37=h2a8f88b_0
  - libssh2=1.9.0=h7a1dbc1_1
  - libtiff=4.2.0=hd0e1b90_0
  - libuv=1.40.0=he774522_0
  - libwebp=1.2.0=h2bbff1b_0
  - llvmlite=0.37.0=py37h23ce68f_1
  - lz4-c=1.9.3=h2bbff1b_1
  - markupsafe=2.0.1=py37h2bbff1b_0
  - matplotlib=3.5.0=py37haa95532_0
  - matplotlib-base=3.5.0=py37h6214cd6_0
  - matplotlib-inline=0.1.2=pyhd3eb1b0_2
  - mistune=0.8.4=py37hfa6e2cd_1001
  - mkl=2021.4.0=haa95532_640
  - mkl-service=2.4.0=py37h2bbff1b_0
  - mkl_fft=1.3.1=py37h277e83a_0
  - mkl_random=1.2.2=py37hf11a4ad_0
  - munkres=1.1.4=py_0
  - nbclient=0.5.3=pyhd3eb1b0_0
  - nbconvert=6.1.0=py37haa95532_0
  - nbformat=5.1.3=pyhd3eb1b0_0
  - nest-asyncio=1.5.1=pyhd3eb1b0_0
  - netcdf4=1.5.7=py37hb33177a_0
  - notebook=6.4.8=py37haa95532_0
  - numba=0.54.1=py37hf11a4ad_0
  - numexpr=2.8.1=py37hb80d3ca_0
  - numpy=1.20.3=py37ha4e8547_0
  - numpy-base=1.20.3=py37hc2deb75_0
  - olefile=0.46=py37_0
  - openssl=1.1.1m=h2bbff1b_0
  - packaging=21.3=pyhd3eb1b0_0
  - pandas=1.3.5=py37h6214cd6_0
  - pandocfilters=1.5.0=pyhd3eb1b0_0
  - parso=0.8.3=pyhd3eb1b0_0
  - pickleshare=0.7.5=pyhd3eb1b0_1003
  - pillow=8.4.0=py37hd45dc43_0
  - pip=21.2.4=py37haa95532_0
  - pluggy=1.0.0=py37haa95532_0
  - prometheus_client=0.13.1=pyhd3eb1b0_0
  - prompt-toolkit=3.0.20=pyhd3eb1b0_0
  - prompt_toolkit=3.0.20=hd3eb1b0_0
  - py=1.11.0=pyhd3eb1b0_0
  - pycparser=2.21=pyhd3eb1b0_0
  - pygments=2.11.2=pyhd3eb1b0_0
  - pyopenssl=22.0.0=pyhd3eb1b0_0
  - pyparsing=3.0.4=pyhd3eb1b0_0
  - pyqt=5.9.2=py37hd77b12b_6
  - pyreadline=2.1=py37_1
  - pyrsistent=0.18.0=py37h196d8e1_0
  - pysocks=1.7.1=py37_1
  - pytest=6.2.5=py37haa95532_2
  - pytest-cov=3.0.0=pyhd3eb1b0_0
  - python=3.7.11=h6244533_0
  - python-dateutil=2.8.2=pyhd3eb1b0_0
  - pytorch=1.10.2=py3.7_cpu_0
  - pytorch-mutex=1.0=cpu
  - pytz=2021.3=pyhd3eb1b0_0
  - pywin32=302=py37h827c3e9_1
  - pywinpty=2.0.2=py37h5da7b33_0
  - pyyaml=5.3.1=py37he774522_0
  - pyzmq=22.3.0=py37hd77b12b_2
  - qt=5.9.7=vc14h73c81de_0
  - qtconsole=5.2.2=pyhd3eb1b0_0
  - qtpy=1.11.2=pyhd3eb1b0_0
  - requests=2.27.1=pyhd3eb1b0_0
  - ruamel.yaml=0.16.12=py37he774522_1
  - ruamel.yaml.clib=0.2.2=py37he774522_0
  - scipy=1.7.3=py37h0a974cb_0
  - send2trash=1.8.0=pyhd3eb1b0_1
  - setuptools=58.0.4=py37haa95532_0
  - sip=4.19.13=py37hd77b12b_0
  - six=1.16.0=pyhd3eb1b0_1
  - snowballstemmer=2.2.0=pyhd3eb1b0_0
  - sphinx=4.4.0=pyhd3eb1b0_0
  - sphinxcontrib-applehelp=1.0.2=pyhd3eb1b0_0
  - sphinxcontrib-devhelp=1.0.2=pyhd3eb1b0_0
  - sphinxcontrib-htmlhelp=2.0.0=pyhd3eb1b0_0
  - sphinxcontrib-jsmath=1.0.1=pyhd3eb1b0_0
  - sphinxcontrib-qthelp=1.0.3=pyhd3eb1b0_0
  - sphinxcontrib-serializinghtml=1.1.5=pyhd3eb1b0_0
  - sqlite=3.37.2=h2bbff1b_0
  - tbb=2021.5.0=h59b6b97_0
  - terminado=0.13.1=py37haa95532_0
  - testpath=0.5.0=pyhd3eb1b0_0
  - tk=8.6.7=vc14hb68737d_1
  - toml=0.10.2=pyhd3eb1b0_0
  - tornado=6.1=py37h2bbff1b_0
  - tqdm=4.62.3=pyhd3eb1b0_1
  - traitlets=5.1.1=pyhd3eb1b0_0
  - typing-extensions=3.10.0.2=hd3eb1b0_0
  - typing_extensions=3.10.0.2=pyh06a4308_0
  - urllib3=1.26.8=pyhd3eb1b0_0
  - vc=14.2=h21ff451_1
  - vs2015_runtime=14.27.29016=h5e58377_2
  - wcwidth=0.2.5=pyhd3eb1b0_0
  - webencodings=0.5.1=py37_1
  - wheel=0.37.1=pyhd3eb1b0_0
  - widgetsnbextension=3.5.2=py37haa95532_0
  - win_inet_pton=1.1.0=py37haa95532_0
  - wincertstore=0.2=py37haa95532_2
  - winpty=0.4.3=4
  - xarray=0.20.1=pyhd3eb1b0_1
  - xz=5.2.5=h62dcd97_0
  - yaml=0.1.7=vc14h4cb57cf_1
  - yapf=0.31.0=pyhd3eb1b0_0
  - zipp=3.7.0=pyhd3eb1b0_0
  - zlib=1.2.11=vc14h1cdd9ab_1
  - zstd=1.4.9=h19a0ad4_0
  - pip:
    - absl-py==1.0.0
    - cachetools==5.0.0
    - google-auth==2.6.0
    - google-auth-oauthlib==0.4.6
    - grpcio==1.43.0
    - markdown==3.3.6
    - nbsphinx==0.8.8
    - nbsphinx-link==1.3.0
    - oauthlib==3.2.0
    - protobuf==3.19.4
    - pyasn1==0.4.8
    - pyasn1-modules==0.2.8
    - requests-oauthlib==1.3.1
    - rsa==4.8
    - sphinx-rtd-theme==1.0.0
    - tensorboard==2.8.0
    - tensorboard-data-server==0.6.1
    - tensorboard-plugin-wit==1.8.1
    - werkzeug==2.0.3
prefix: C:\Users\321319\Miniconda3\envs\neuralhydrology
```
