### 2022-02-23 neuralhydrology/neuralhydrology#77 @chuckaustin
[REQUEST] A bit more detail on generic datasets 
[Part of JOSS Review: https://github.com/https://github.com/https://github.com/https://github.com/openjournals/joss-reviews/issues/4050]

I've enjoyed going through this software as part of its JOSS review, and I am excited to use it in my own research, possibly quite soon.

One final issue I'd like to bring up is a request for more information on generic datasets in Example 3. I know the tutorials can't cover every detail of the software, but I think it would be really handy to see an example of data that has been formatted in a different way than the CAMELS dataset and imported as a GenericDataset object -- I see myself (and probably many other users) doing this in the future.

One detail that would be useful to know, for example, is whether the software expects to see missing data values entered in a certain way (are they flagged? entered as -9999? etc...). It would also be helpful to know how the code handles missing data during training (i.e., are time steps without real data automatically excluded from the objective function?) Having this explained up front in the tutorial, and even better, showing an example of importing data into a generic dataset, would be much appreciated.
