### 2022-02-23 #4050 @elbeejay
Hello all :wave:

I wanted to thank our reviewers for testing and evaluating this software. To hopefully help @kratzert and co-authors make changes, I wanted to ask @jhamman to provide some clarification on the unchecked boxes in the documentation and software paper sections. Please either open issues in the repositories or make a small comment here if these are concerns about missing documentation, tests, or community guidelines. I believe the paper has been revised since your review to try and address some of your concerns.

Similarly @chuckaustin, if you could check off the boxes corresponding to things you believe the authors have accomplished, this would help narrow the scope of the revisions you'd like to see made, and complement the open issues.

Again thanks all, @kratzert please let me or any of the reviewers know if you need further clarification about open issues or unchecked items on reviewer checklists.

Thanks,
Jay
