### 2022-02-28 neuralhydrology/neuralhydrology#77 @kratzert
Hi @chuckaustin

good suggestion. I will make sure that we add a more detailed description on how to work with the GenericDataset in our next update. Also sorry for the late reply, I was on a business trip the last week and hadn't had much time to look at Github.
