### 2022-02-28 neuralhydrology/neuralhydrology#76 @kratzert
Hi @chuckaustin

I only know this problem if a) you installed the CPU version of PyTorch or b) PyTorch has problems detecting your GPU. Since you said you used the cuda 9.2 environment file of our library my only guess is that it has to be (b). As far as I could see from the NVIDIA homepage, your GPU should support this CUDA version. Are you sure you have an updated NVIDIA driver installed? Usually there is nothing more to do from the client side (e.g. manually installing CUDA or cudNN is not required).

However, I fear that this is not really an issue with our library and I don't know what we could do to fix this.
