### 2022-02-28 neuralhydrology/neuralhydrology#76 @kratzert
Okay, I had another look at it. From the `conda env export` output you posted in #75  I can see that PyTorch did not install the corresponding cuda version but still the cpu version (see: `pytorch=1.10.2=py3.7_cpu_0`). I am not quite sure why this happens. I tested what happens if I create a conda env from the same environment file on my machine and I get exactly the same. Using the `environment_cuda10_2.yml` file however, I get a correctly working PyTorch installation with GPU support. I am not sure if they maybe dropped the CUDA 9.2 support.

From [here](https://pytorch.org/get-started/previous-versions/) It seems like that PyTorch 1.7.1 was the last version with CUDA 9.2 support, however our environment files seem to install PyTorch 1.10.2, which only works with CUDA 10.2 or 11.3 (see [here](https://pytorch.org/get-started/locally/)). If your GPU supports CUDA 10.2 do you mind creating and environment from the other file and testing if it works then? Seems like we should remove the 9.2 environment file.
If your GPU does not support CUDA 10.2, you could install the old PyTorch version with CUDA 9.2 support by:

```
conda install pytorch==1.7.1 cudatoolkit=9.2 -c pytorch
```
As far as I know, we do not use any features that were added recently to PyTorch so the old version should work equally fine.
