### 2022-02-28 #4050 @kratzert
Hi @elbeejay

sorry for my delayed reply. I was on a business trip and hadn't had much time to look into it in the last days. I just checked all remaining review comments in our repository and I think they are pretty clear.

- I think I could solve the installation issues of @chuckaustin but I am waiting for his confirmation (see https://github.com/neuralhydrology/neuralhydrology/issues/76#issuecomment-1054656670). We will remove the CUDA9.2 environment file and we will add one for cuda11.3 (which, together with 10.2, are the two supported CUDA versions for the stable PyTorch installation).
- I will also update the `GenericDataset` description in the tutorial, but I am waiting for feedback of @chuckaustin before committing, see https://github.com/neuralhydrology/neuralhydrology/issues/77#issuecomment-1054636471

Other than that, I think there is nothing left open for me/us to do, right? I am not sure about the unchecked checkboxes above though.
