### 2022-03-01 #4050 @elbeejay
Sounds good @kratzert, it looks like the installation and example checkboxes from @chuckaustin's review are all that remain. It seems like the path to resolving the two outstanding issues is relatively clear which is good to hear. You are correct that there is nothing else for you and your co-authors to do at this time.

Once the outstanding issues are resolved, I'll do a final check of the submitted materials, ask you to create an archive and tagged release for the reviewed code, and that's about it.

Thanks,
Jay
