### 2022-03-01 neuralhydrology/neuralhydrology#77 @chuckaustin
Hi @kratzert, thank you for your explanation, it's much appreciated. You addressed the questions I was curious about. As you said, if the first point at least could be added to the tutorials in the future, I think it would be very helpful for new users.
