### 2022-03-01 neuralhydrology/neuralhydrology#76 @chuckaustin
Hi @kratzert thank you for your suggestion, deleting the environment then creating it using` environment_cuda10_2.yml` got it running on my GPU, which made the training much faster. I did notice that, even on the GPU, training was much faster when I set "num_workers" to 0 instead of 8.
