### 2022-03-02 e5b3bd9
Remove cuda9.2 env (#137)
* Removed cuda9.2 env, added cuda11.3 env

* Renamed env
