### 2022-03-03 #4050 @elbeejay
Hi @kratzert, it looks like those last few issues have been resolved. I wanted to follow up on some of the discussion that happened in [issue 77](https://github.com/neuralhydrology/neuralhydrology/issues/77) - it sounded like you and @chuckaustin agreed that it would be useful to expand the documentation in response to some of the questions raised in that issue. Do you plan on making those changes to the documentation prior to publication? I see that the issue was closed, so this is by no means a requirement for publication, but I figured I would ask first.

Thanks
