### 2022-03-03 #4050 @elbeejay
@kratzert one minor suggestion for the paper text:
- Line 19: Switch the colon to a period, or remove the capital "P" in "Pre-built" on the following line

Besides this, I think things look good. I am optimistic that this package will help accelerate deep learning acceptance/adoption in the hydrology community. After making the minor revision noted above, I need you to do a few things at this time:

Please let me know when these steps are complete, they allow us to finalize the metadata that will be associated with the JOSS publication. Once that is done I can recommend that we accept and publish this paper.
