### 2022-03-03 15d4e6b
update description/doc of GenericDataset (#138)
* update description/doc of GenericDataset

* Update neuralhydrology/datasetzoo/genericdataset.py

Co-authored-by: Martin Gauch <15731649+gauchm@users.noreply.github.com>

Co-authored-by: Martin Gauch <15731649+gauchm@users.noreply.github.com>
