### 2022-03-03 #4050 @kratzert
> @kratzert one minor suggestion for the paper text:
>
>     * Line 19: Switch the colon to a period, or remove the capital "P" in "Pre-built" on the following line
>
>
> Besides this, I think things look good. I am optimistic that this package will help accelerate deep learning acceptance/adoption in the hydrology community. After making the minor revision noted above, I need you to do a few things at this time:
>
>     * [ ]  Make a tagged release of your software, and list the version tag of the archived version here.
>
>     * [ ]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
>
>     * [ ]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
>
>     * [ ]  Please list the DOI of the archived version here.
>
>
> Please let me know when these steps are complete, they allow us to finalize the metadata that will be associated with the JOSS publication. Once that is done I can recommend that we accept and publish this paper.

- Link to tagged version: https://github.com/neuralhydrology/neuralhydrology/releases/tag/v.1.2.3
- Link to archived software: https://zenodo.org/record/6326394
- Checked all authors names, added ORCID ids
- DOI: https://doi.org/10.5281/zenodo.6326394

The paper is also updated with the typo fix.

Anything else you need?
