### 2022-03-04 #4050 @kratzert
> > Anything else you need?
>
> @kratzert that is it, thank you for taking care of those last items. I wanted to thank the folks that peer-reviewed this submission once again as well - so thank you to @jhamman, @ammilten, and @chuckaustin.

Also thank you from my side for taking care of the entire review process @elbeejay and  @jhamman, @ammilten and @chuckaustin  for reviewing our package and providing useful feedback!

> At this time I will be recommending that we accept and publish this submission in JOSS. The paper and metadata will get one final review by an Editor-in-Chief prior to publication.

Juhu :tada:
