### 2022-03-04 #4050 @kyleniemeyer
Hi @kratzert, I made a small PR with some formatting fixes to the paper: https://github.com/neuralhydrology/neuralhydrology/pull/79

Could you merge this? Then I can publish the paper.
