### 2022-03-04 #4050 @kratzert
> Hi @kratzert, I made a small PR with some formatting fixes to the paper: [neuralhydrology/neuralhydrology#79](https://github.com/neuralhydrology/neuralhydrology/pull/79)
>
> Could you merge this? Then I can publish the paper.

Done. Thanks for the fixes.
